# Conflicts:
#	.classpath
#	WebContent/Adm/visualizarLivro.xhtml
#	build/classes/br/com/fiap/si/modelo/Categoria.class

+++ src/br/com/fiap/si/dao/LivroDAO.java
@@ -15,5 +15,5 @@ public interface LivroDAO {
 	
 	public List<Livro> selectLivro();
 	
-	public Livro getLivroID(Long id);
+	public Livro getLivroID(Integer id);
 }
